+++ inst/plumber/plumber.R
@@ -713,15 +713,17 @@ function(ejamitout = testoutput_ejamit_10pts_1miles, ..., attachment = "true", r
   #   `Age` = c('pctunder5', 'pctunder18', 'pctover64'),
   #   `Community` = names_community[!(names_community %in% c( 'pctmale', 'pctfemale', 'pctownedunits_dupe'))],
   #   `Poverty` = names_d_extra,
-  #   `Features and Location Information` = c(
-  #     names_e_other,
-  #     names_sitesinarea,
+  #   `Counts of Features and Overlap with Area Types` = c(
   #     names_featuresinarea,
   #     names_flag
   #   ),
+  #   `Critical Services` = c('yesno_houseburden', 'yesno_fooddesert', 'yesno_transdis',
+  #                           'pctnobroadband', 'pctnohealthinsurance'),
   #   `Climate` = names_climate,
-  #   `Critical Services` = names_criticalservice,
-  #   `Other` = names_d_other_count
+  #   `Other Totals` = names_d_other_count,
+  #   `Facility Counts` = names_sitesinarea,
+  #   `Analyzed Sites` = c('distance_min_avgperson', 'distance_min',
+  #                        'sitecount_unique', 'sitecount_avg', 'sitecount_max')
   #   # , `Count above threshold` = names_countabove  # need to fix map_headernames longname and calctype and weight and drop 2 of the 6
   # ),
   # ## all the indicators that are in extratable_list_of_sections: